No more go-horse/experimental code
Made changes into a more definite state of the code. Excluded lots of
dupes made due to experimental code and rearranged classes and methods.
