+++ .gitignore
@@ -45,3 +45,7 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
+OverwatcherWpf/OverwatcherWpf/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
+*.txt
+OverwatcherWpf/OverwatcherWpf/obj/Debug/OverwatcherWpf_MarkupCompile.i.cache
+*.cache
