+++ OverwatcherWpf/OverwatcherWpf/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace OverwatcherWpf
+namespace Overwatcher
 {
     /// <summary>
     /// Interaction logic for App.xaml
